Add MoM and YoY on summary page

+++ tests/stepper.rs
@@ -16,10 +16,10 @@ impl FieldStepper for Testing {}
 impl DataVerifier for Testing {}
 
 fn create_test_db(file_name: &str) -> Connection {
-    if let Ok(metadata) = fs::metadata(file_name) {
-        if metadata.is_file() {
-            fs::remove_file(file_name).expect("Failed to delete existing file");
-        }
+    if let Ok(metadata) = fs::metadata(file_name)
+        && metadata.is_file()
+    {
+        fs::remove_file(file_name).expect("Failed to delete existing file");
     }
 
     let mut conn = Connection::open(file_name).unwrap();
